// src/types/logs.ts
export type Logs = {
    dados: {

        id: string;
        accao: string;
        recurso_afetado: string;
        valor: string;
        justificativa: string | null;
        ip_address: string;
        canal: string;
        cargo: string;
        user_name: string;
        created_at: string;
        updated_at: string;
        status: string

    }
}